[refactor] 도메인 별로 패키지 분리
