+++ src/main/java/com/soolsul/soolsulserver/domain/curation/CurationPicture.java
